Updated to core 6.0.0-unstable1554
