+++ src/NServiceBus.Callbacks.Tests/packages.config
@@ -4,6 +4,6 @@
   <package id="ApprovalTests" version="3.0.10" targetFramework="net452" />
   <package id="ApprovalUtilities" version="3.0.10" targetFramework="net452" />
   <package id="Mono.Cecil" version="0.9.6.1" targetFramework="net45" />
-  <package id="NServiceBus" version="6.0.0-unstable1543" targetFramework="net452" />
+  <package id="NServiceBus" version="6.0.0-unstable1554" targetFramework="net452" />
   <package id="NUnit" version="2.6.4" targetFramework="net45" />
 </packages>
