+++ src/NServiceBus.Callbacks/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="GitVersionTask" version="2.0.1" targetFramework="net45" developmentDependency="true" />
-  <package id="NServiceBus" version="6.0.0-unstable1543" targetFramework="net452" />
+  <package id="NServiceBus" version="6.0.0-unstable1554" targetFramework="net452" />
   <package id="NuGetPackager" version="0.5.5" targetFramework="net451" developmentDependency="true" />
 </packages>
